projeto03/requirements.txt
python-dotenv
openai
groq
PyPDF2
python-docx
